Restructure the components directory inside apps/super-app
