--- apps/super-app/src/common/components/index.ts
@@ -1,1 +1,0 @@
-export { default as Container } from './Container';
