+++ apps/super-app/src/common/components/index.tsx
@@ -1,0 +1,3 @@
+export { default as Container } from './Container';
+export { SupportDetailPage } from './support-detail-page/SupportDetailPage';
+export { ServiceCard } from './service-card/ServiceCard';
