+++ apps/super-app/src/common/components/support-detail-page/SupportDetailPage.css
