--- apps/super-app/src/components/index.ts
@@ -1,2 +1,0 @@
-export { NavBar } from './navbar';
-export { NavBarContainer } from './NavBarContainer';
