--- apps/super-app/src/routes/index.tsx
@@ -1,3 +1,0 @@
-export { LandingPage } from './LandingPage';
-export { Options } from './Options';
-export { SupportDetailPage } from './SupportDetailPage';
